pyproject.toml
[project]
name = "llm-router"
version = "1.0.0"
description = "A unified API router for multiple language model providers"
authors = [{ name = "LLM Router", email = "router@example.com" }]
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "pydantic>=2.0.0",
  "pyyaml>=6.0.0",
  "requests>=2.31.0",
  "aiohttp>=3.9.0",
  "langchain>=0.2.0,<0.3.0",
  "langchain-openai>=0.1.7",
  "langchain-anthropic>=0.1.0",
  "langchain-community>=0.2.0,<0.3.0",
  "openai>=1.3.0",
  "anthropic>=0.7.0",
  "torch>=2.0.0",
  "transformers>=4.51.1,<4.52.0",
  "accelerate>=0.24.0",
  "bitsandbytes>=0.46.1",
  "pillow>=10.0.0",
  "vllm==0.8.5",
  "tiktoken>=0.7.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = [
    "src",
    "pyproject.toml",
    "README.md",
    "configs",
    "test_lazy_loading.py",
    "test_routing.py",
    "test_routing.sh",
]

[project.scripts]
llm-router = "src.main:main"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.black]
line-length = 100
target-version = ['py312']
